Refactor Dru::Command#run_docker_compose_command
Instead of using a custom tty option to decide to run `system` or
`TTY:Command` when we need to attach to the stdin, we add a custom run
method that sets an option to `TTY::Command` to always use `/dev/tty` as
stdin.
`Dru::Command#run_docker_compose_command` now uses this method.

Changed also the `#container_name_to_id` method to use the new
`#run_docker_compose_command` and removed the default `container_name`.